feat: 调整int类型为long型

+++ boot/authority-web-boot-starter/src/main/java/com/kewen/framework/boot/authority/biz/service/impl/MemorySysMenuAuthUnify.java
@@ -1,7 +1,7 @@
 package com.kewen.framework.boot.authority.biz.service.impl;
 
-import com.kewen.framework.boot.authority.biz.mapper.entity.SysMenu;
-import com.kewen.framework.boot.authority.biz.mapper.entity.SysMenuAuth;
+import com.kewen.framework.boot.authority.biz.entity.SysMenu;
+import com.kewen.framework.boot.authority.biz.entity.SysMenuAuth;
 import com.kewen.framework.boot.authority.biz.model.resp.MenuResp;
 import com.kewen.framework.boot.authority.biz.service.SysMenuAuthService;
 import com.kewen.framework.boot.authority.biz.service.SysMenuAuthUnify;
@@ -60,7 +60,7 @@ public class MemorySysMenuAuthUnify implements SysMenuAuthUnify {
     @Override
     public List<MenuResp> getMenuTree() {
         List<SysMenu> sysMenus = getSysMenus();
-        Map<Integer, List<SysMenuAuth>> authByMenuMap = getSysMenuAuths().stream()
+        Map<Long, List<SysMenuAuth>> authByMenuMap = getSysMenuAuths().stream()
                 .collect(Collectors.groupingBy(SysMenuAuth::getMenuId));
         List<MenuResp> collect = sysMenus.stream()
                 .map(l -> BeanUtil.toBean(l, MenuResp.class))
@@ -132,13 +132,13 @@ public class MemorySysMenuAuthUnify implements SysMenuAuthUnify {
     private boolean hasMenuAuth(Collection<String> authorities, SysMenu sysMenu){
         //基于自己的权限
         if (MenuAuthType.SELF==sysMenu.getAuthType()){
-            Integer menuId = sysMenu.getId();
+            Long menuId = sysMenu.getId();
             return getSysMenuAuths().stream()
                     .filter(a -> a.getMenuId().equals(menuId))
                     .anyMatch(a -> authorities.contains(a.getAuthority()));
         } else {
             //基于父菜单的权限
-            Integer parentId = sysMenu.getParentId();
+            Long parentId = sysMenu.getParentId();
             Optional<SysMenu> first = getSysMenus().stream().filter(m -> m.getId().equals(parentId)).findFirst();
             if (!first.isPresent()){
                 //已经到顶了，找不到 top对应的parent的菜单
